Change version numbers to long

+++ service/src/main/java/hu/rbr/sfinapp/account/AccountService.java
@@ -48,7 +48,7 @@ public class AccountService extends BaseService implements Versioned {
     }
 
     @Override
-    public int getVersion() {
+    public long getVersion() {
         return version.get();
     }
