+++ service/src/main/java/hu/rbr/sfinapp/core/service/Versioned.java
@@ -2,6 +2,6 @@ package hu.rbr.sfinapp.core.service;
 
 public interface Versioned {
 
-    int getVersion();
+    long getVersion();
 
 }
